+++ src/pyimpspec/data/formats/crv.py
@@ -1,0 +1,104 @@
+# pyimpspec is licensed under the GPLv3 or later (https://www.gnu.org/licenses/gpl-3.0.html).
+# Copyright 2026 pyimpspec developers
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+#
+# The licenses of pyimpspec's dependencies and/or sources of portions of code are included in
+# the LICENSES folder.
+
+from pyimpspec.data.data_set import (
+    DataSet,
+    dataframe_to_data_sets,
+)
+from pyimpspec.exceptions import UnsupportedFileFormat
+from pyimpspec.typing.helpers import (
+    IO,
+    List,
+    Path,
+    Union,
+)
+from .helpers import (
+    _parse_string_as_float,
+    _validate_path,
+)
+
+
+def parse_crv(path: Union[str, Path]) -> List[DataSet]:
+    """
+    Parse an OrigaLys .crv file containing an impedance spectrum.
+
+    Parameters
+    ----------
+    path: Union[str, pathlib.Path]
+        The path to the file to process.
+
+    Returns
+    -------
+    List[DataSet]
+    """
+    from pandas import DataFrame
+
+    _validate_path(path)
+
+    fp: IO
+    with open(path, "r", encoding="latin1") as fp:
+        lines: List[str] = list(
+            filter(lambda _: _ != "", map(str.strip, fp.readlines()))
+        )
+
+    while lines:
+        line: str = lines.pop(0)
+        if "EIS" in line:
+            break
+    else:
+        raise UnsupportedFileFormat("Expected to find a header line containing 'EIS'")
+
+    num_separators: int = 8
+
+    while lines and lines[0].count("\t") < num_separators:
+        line = lines.pop(0)
+
+    if not line.isnumeric():
+        raise UnsupportedFileFormat(
+            f"Expected a line that specifies the number of data points instead of {line=}"
+        )
+
+    num_points: int = int(line)
+
+    if not lines:
+        raise UnsupportedFileFormat(f"Expected to fine lines containing at least five columns instead of {line=}")
+
+    freq: List[float] = []
+    real: List[float] = []
+    imag: List[float] = []
+
+    while lines and lines[0].count("\t") == num_separators:
+        values = list(map(_parse_string_as_float, lines.pop(0).split("\t")))
+        freq.append(values[0])
+        real.append(values[6])
+        imag.append(-values[7])
+
+    if len(freq) != num_points:
+        raise ValueError(f"Expected to parse {num_points} data points instead of {len(freq)}")
+
+    return dataframe_to_data_sets(
+        DataFrame.from_dict(
+            {
+                "frequency": freq,
+                "real": real,
+                "imaginary": imag,
+            }
+        ),
+        path=path,
+    )